+++ changes.md
@@ -1,10 +1,11 @@
 ## Main wip
 
 ## 2.0.3
-* Bit by windows file case insensitive files
+* Added industrial pipe annotations
+* Added industrial stair annotations
   
 ## 2.0.2
-* Added industrial pipe annotations
+* Bit by windows file case insensitive files
   
 ## 2.0.1
 * Cleaned up IndustrialStairs example